[SYCL][UR] Combined CTS fixes for CL adapter. (#11794)
Pulls in fixes from
https://github.com/oneapi-src/unified-runtime/pull/1044

+++ sycl/source/detail/context_impl.cpp
@@ -77,7 +77,7 @@ context_impl::context_impl(sycl::detail::pi::PiContext PiContext,
       MSupportBufferLocationByDevices(NotChecked) {
 
   std::vector<sycl::detail::pi::PiDevice> DeviceIds;
-  size_t DevicesNum = 0;
+  uint32_t DevicesNum = 0;
   // TODO catch an exception and put it to list of asynchronous exceptions
   Plugin->call<PiApiKind::piContextGetInfo>(
       MContext, PI_CONTEXT_INFO_NUM_DEVICES, sizeof(DevicesNum), &DevicesNum,
